Network Subscriptions; Move all subscription tests to separate subscription_test file
