+++ dh/network.go
@@ -2,6 +2,8 @@ package dh
 
 import (
 	"encoding/json"
+	"strconv"
+	"time"
 )
 
 type Network struct {
@@ -86,4 +88,35 @@ func (c *Client) ListNetworks(params *ListParams) (list []*Network, err *Error) 
 	}
 
 	return list, nil
+}
+
+func (n *Network) SubscribeInsertCommands(names []string, timestamp time.Time) (subs *CommandSubscription, err *Error) {
+	return n.subscribeCommands(names, timestamp, false)
+}
+
+func (n *Network) SubscribeUpdateCommands(names []string, timestamp time.Time) (subs *CommandSubscription, err *Error) {
+	return n.subscribeCommands(names, timestamp, true)
+}
+
+func (n *Network) subscribeCommands(names []string, timestamp time.Time, isCommUpdatesSubscription bool) (subs *CommandSubscription, err *Error) {
+	id := []string{strconv.FormatInt(n.Id, 10)}
+	params := &SubscribeParams{
+		Names:                 names,
+		Timestamp:             timestamp,
+		ReturnUpdatedCommands: isCommUpdatesSubscription,
+		NetworkIds:            id,
+	}
+
+	return n.client.SubscribeCommands(params)
+}
+
+func (n *Network) SubscribeNotifications(names []string, timestamp time.Time) (subs *NotificationSubscription, err *Error) {
+	id := []string{strconv.FormatInt(n.Id, 10)}
+	params := &SubscribeParams{
+		Names:      names,
+		Timestamp:  timestamp,
+		NetworkIds: id,
+	}
+
+	return n.client.SubscribeNotifications(params)
 }
